refactor: rename bn254_grumpkin -> bn256_grumpkin
